# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MdM Control Backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/inventory-search": {
      "get": {
        "tags": [
          "InventorySearch"
        ],
        "summary": "List Inventory Searches",
        "operationId": "list_inventory_searches_api_v1_inventory_search_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_InventorySearchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "InventorySearch"
        ],
        "summary": "Create Inventory Search",
        "operationId": "create_inventory_search_api_v1_inventory_search_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorySearchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorySearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory-search/{search_id}": {
      "get": {
        "tags": [
          "InventorySearch"
        ],
        "summary": "Get Inventory Search",
        "operationId": "get_inventory_search_api_v1_inventory_search__search_id__get",
        "parameters": [
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Search Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorySearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "InventorySearch"
        ],
        "summary": "Update Inventory Search",
        "operationId": "update_inventory_search_api_v1_inventory_search__search_id__put",
        "parameters": [
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Search Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorySearchUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorySearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "InventorySearch"
        ],
        "summary": "Delete Inventory Search",
        "operationId": "delete_inventory_search_api_v1_inventory_search__search_id__delete",
        "parameters": [
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Search Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory-search/execute": {
      "post": {
        "tags": [
          "InventorySearch"
        ],
        "summary": "Execute Inventory Search",
        "operationId": "execute_inventory_search_api_v1_inventory_search_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventorySearchExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  },
                  "type": "array",
                  "title": "Response Execute Inventory Search Api V1 Inventory Search Execute Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile-apps": {
      "get": {
        "tags": [
          "MobileApps"
        ],
        "summary": "List Mobile Apps",
        "operationId": "list_mobile_apps_api_v1_mobile_apps_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_MobileAppResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MobileApps"
        ],
        "summary": "Create Mobile App",
        "operationId": "create_mobile_app_api_v1_mobile_apps_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileAppCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileAppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile-apps/{mobile_app_id}": {
      "get": {
        "tags": [
          "MobileApps"
        ],
        "summary": "Get Mobile App",
        "operationId": "get_mobile_app_api_v1_mobile_apps__mobile_app_id__get",
        "parameters": [
          {
            "name": "mobile_app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mobile App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileAppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MobileApps"
        ],
        "summary": "Update Mobile App",
        "operationId": "update_mobile_app_api_v1_mobile_apps__mobile_app_id__put",
        "parameters": [
          {
            "name": "mobile_app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mobile App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileAppUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileAppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MobileApps"
        ],
        "summary": "Delete Mobile App",
        "operationId": "delete_mobile_app_api_v1_mobile_apps__mobile_app_id__delete",
        "parameters": [
          {
            "name": "mobile_app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mobile App Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Certificate": {
        "properties": {
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "expiry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          }
        },
        "type": "object",
        "title": "Certificate"
      },
...
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "Online",
          "Offline",
          "Pending"
        ],
        "title": "ConnectionStatus"
      },
      "Criteria": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "operator": {
            "type": "string",
            "title": "Operator"
          },
          "type": {
            "$ref": "#/components/schemas/CriteriaType"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "and_or": {
            "type": "string",
            "title": "And Or",
            "default": "AND"
          },
          "left_parentheses": {
            "type": "boolean",
            "title": "Left Parentheses",
            "default": false
          },
          "right_parentheses": {
            "type": "boolean",
            "title": "Right Parentheses",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "field",
          "operator",
          "type",
          "value"
        ],
        "title": "Criteria"
      },
      "CriteriaType": {
        "type": "string",
        "enum": [
          "string",
          "number",
          "boolean",
          "date"
        ],
        "title": "CriteriaType"
      },
...
      "DeviceResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number"
          },
          "os_version": {
            "type": "string",
            "title": "Os Version"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "enrollment_status": {
            "$ref": "#/components/schemas/EnrollmentStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_enrolled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Enrolled At"
          },
          "battery_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Status"
          },
          "total_storage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Storage"
          },
          "available_storage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Storage"
          },
          "total_memory": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Memory"
          },
          "available_memory": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Memory"
          },
          "network": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Network"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network"
          },
          "certificates": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Certificate"
                },
                "type": "array"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificates"
          },
          "extension_attributes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DeviceExtensionAttributeResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Attributes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "serial_number",
          "os_version",
          "connection_status",
          "enrollment_status",
          "created_at",
          "updated_at",
          "last_enrolled_at"
        ],
        "title": "DeviceResponse"
      },
...
      "EnrollmentStatus": {
        "type": "string",
        "enum": [
          "Compliant",
          "Non-compliant",
          "Needs attention",
          "Enrolled",
          "Pending",
          "Unknown"
        ],
        "title": "EnrollmentStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventorySearchCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/Criteria"
            },
            "type": "array",
            "minItems": 1,
            "title": "Criteria"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "name",
          "criteria",
          "created_by"
        ],
        "title": "InventorySearchCreate"
      },
      "InventorySearchExecuteRequest": {
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/Criteria"
            },
            "type": "array",
            "minItems": 1,
            "title": "Criteria"
          }
        },
        "type": "object",
        "required": [
          "criteria"
        ],
        "title": "InventorySearchExecuteRequest"
      },
      "InventorySearchResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/Criteria"
            },
            "type": "array",
            "title": "Criteria",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "created_by",
          "updated_at"
        ],
        "title": "InventorySearchResponse"
      },
      "InventorySearchUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "criteria": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Criteria"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria"
          }
        },
        "type": "object",
        "title": "InventorySearchUpdate"
      },
      "MobileAppCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "package_name": {
            "type": "string",
            "title": "Package Name"
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          }
        },
        "type": "object",
        "required": [
          "name",
          "enabled",
          "version",
          "package_name",
          "scope"
        ],
        "title": "MobileAppCreate"
      },
      "MobileAppResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "package_name": {
            "type": "string",
            "title": "Package Name"
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "enabled",
          "version",
          "package_name",
          "scope"
        ],
        "title": "MobileAppResponse"
      },
      "MobileAppUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Scope"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "MobileAppUpdate"
      },
      "Network": {
        "properties": {
          "wifi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Wifi"
              },
              {
                "type": "null"
              }
            ]
          },
          "cellular": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Cellular"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "Network"
      },
...
      "PaginatedResponse_InventorySearchResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InventorySearchResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PaginatedResponse[InventorySearchResponse]"
      },
      "PaginatedResponse_MobileAppResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MobileAppResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PaginatedResponse[MobileAppResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}